MA-19378 stop building when fail to execute target mxmwifi recipe
Previously, the $(MAKE) command in the recipe of target mxmwifi is
passed to shell with the cp command together, if the $(MAKE) command
fails but the cp command succeed, the building process won't stop.

In this patch, pass the $(MAKE) and cp commands individually to shell,
then any failure will cause the build process to stop.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: Id22642018e20620cd021cdd9fc79a2f60202e9da