MA-19378-1 force to update the shell env script
There is a recipe with its target being a shell script. This shell
script is redirected with export shell env commands and is executed when
build the module.

It is possible that the shell scripts already exists but commands
redirected to it need to be changed. To eliminate build errors, force to
update this shell env script by make it a phony target.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I700a8025d0ca78bbaf00eb9206ff3a6cccbf9a33