mxm_wifiex: fix the build errors with the API changes on next-20220719 kernel tag
There are many cfg80211 common api and struct changes on next-20220719
kernel tag, which is based on v5.19-rc7 Linus tree tag with the for-next
branch patches on each kernel IP repo, target for v6.0 kernel version.
So need to change the local wifi driver accordingly to avoid the build
errors.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>